+++ tests/AutoMapperTest.php
@@ -42,6 +42,8 @@ use AutoMapper\Tests\Fixtures\HasDateTimeWithNullValue;
 use AutoMapper\Tests\Fixtures\Issue111\Colour;
 use AutoMapper\Tests\Fixtures\Issue111\ColourTransformer;
 use AutoMapper\Tests\Fixtures\Issue111\FooDto;
+use AutoMapper\Tests\Fixtures\Issue189\User as Issue189User;
+use AutoMapper\Tests\Fixtures\Issue189\UserPatchInput as Issue189UserPatchInput;
 use AutoMapper\Tests\Fixtures\ObjectsUnion\Bar;
 use AutoMapper\Tests\Fixtures\ObjectsUnion\Foo;
 use AutoMapper\Tests\Fixtures\ObjectsUnion\ObjectsUnionProperty;
@@ -1009,7 +1011,7 @@ class AutoMapperTest extends AutoMapperBaseTest
         $input = new Fixtures\SkipNullValues\Input();
 
         /** @var Fixtures\SkipNullValues\Entity $entity */
-        $entity = $this->autoMapper->map($input, $entity, ['skip_null_values' => true]);
+        $entity = $this->autoMapper->map($input, $entity, [MapperContext::SKIP_NULL_VALUES => true]);
         self::assertEquals('foobar', $entity->getName());
     }
 
@@ -1353,7 +1355,7 @@ class AutoMapperTest extends AutoMapperBaseTest
 
         self::assertSame(
             ['bar' => 'bar'],
-            $this->autoMapper->map(new Uninitialized(), 'array', ['skip_null_values' => true])
+            $this->autoMapper->map(new Uninitialized(), 'array', [MapperContext::SKIP_UNINITIALIZED_VALUES => true])
         );
     }
 
@@ -1602,5 +1604,19 @@ class AutoMapperTest extends AutoMapperBaseTest
             'bar' => 'bar',
             'foo' => ['foo1', 'foo2'],
         ], $array);
+    }
+
+    public function testUninitializedProperties(): void
+    {
+        $payload = new Issue189UserPatchInput();
+        $payload->firstName = 'John';
+        $payload->lastName = 'Doe';
+
+        /** @var Issue189User $data */
+        $data = $this->autoMapper->map($payload, Issue189User::class, [MapperContext::SKIP_UNINITIALIZED_VALUES => true]);
+
+        $this->assertEquals('John', $data->getFirstName());
+        $this->assertEquals('Doe', $data->getLastName());
+        $this->assertTrue(!isset($data->birthDate));
     }
 }